medipass-python-backend/requirements.txt
flask>=2.0
openai>=1.0.0
llama-cpp-python>=0.1.78